requirements.txt
requests
beautifulsoup4
streamlit
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
tqdm
